pyproject.toml
[project]
name = "fastactor"
version = "0.1.0"
description = "FastActor is a fast and easy-to-use actor framework for Python."
authors = [{ name = "Cyrus Nouroozi", email = "cyrus@zenbase.ai" }]
dependencies = [
    "anyio>=4.7.0",
    "pyee>=12.1.1",
    "pydantic-settings>=2.7.0",
    "sorcery>=0.2.2",
    "svix-ksuid>=0.6.2",
    "beartype>=0.19.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.23.0",
    "ipdb>=0.13.13",
    "pytest>=8.3.3",
    "pytest-anyio>=0.0.0",
    "trio>=0.25.0",
    "pydantic>=2.10.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastactor"]
